+++ sharedPlugins/clusterioMod/dole_nn_base.js
