+++ test/missing.js
@@ -2,7 +2,6 @@
 // required here in order to track coverage for them.
 
 require('lib/clusterTools');
-require('lib/dole_nn_base');
 require('lib/getInstances');
 require('lib/manager/instanceManager');
 require('lib/manager/modManager');
