+++ views/index.ejs
@@ -1,7 +1,6 @@
 <!doctype HTML>
 <head>
     <%- include("head") %>
-    <script defer src="storage.js"></script>
 </head>
 
 <body>
@@ -10,40 +9,9 @@
     <div id="storage">
         <div class="page-header mt-2 row" id="banner">
             <div class="row">
-                <div class="col-lg-8 col-md-7 col-sm-6">
-                    <h2>Storage</h2>
-                </div>
+                <h2>Work in progress interface</h2>
             </div>
-        </div>
-        <div class="form-group mt-3 mb-3 row">
-            <input type="text" class="form-control" id="search" placeholder="search">
-        </div>
-        <div id="contents" class="row w-100">
-            <table class="table table-striped table-hover">
-                <thead class="thead-dark">
-                <tr>
-                    <th scope="col">Icon</th>
-                    <th scope="col">Item ID</th>
-                    <th scope="col">Quantity</th>
-                </tr>
-                </thead>
-                <tbody><!-- populated by js --></tbody>
-            </table>
-        </div>
-    </div>
-    <div id="graphDisplay">
-        <div id="storageGraphContainer">
-            <div id="contentGraph" class="storageGraph" style="width: 100%;"></div>
         </div>
     </div>
 </div>
-<script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"
-        integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo"
-        crossorigin="anonymous"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js"
-        integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49"
-        crossorigin="anonymous"></script>
-<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js"
-        integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy"
-        crossorigin="anonymous"></script>
 </body>
