+++ views/nav.ejs
@@ -6,9 +6,6 @@
         </a>
     </div>
     <div class="nav-item mr-1">
-        <a class="nav-link align-middle" href="/">Storage</a>
-    </div>
-    <div class="nav-item mr-1">
         <a class="nav-link align-middle" href="/nodes">Slaves</a>
     </div><%
 for(let o in locals.masterPlugins) {
